one query per file
